ultimos cambios

+++ src/data/inmuebles.js
@@ -1,6 +1,6 @@
 export async function fetchObtenerInmuebles(){
 
-    const respuesta = await fetch(import.meta.env.VITE_API_URL)
+    const respuesta = await fetch(`${import.meta.env.VITE_API_URL}/fetchInmuebles`)
     const resultado = await respuesta.json()
     return resultado
 }
@@ -8,14 +8,14 @@ export async function fetchObtenerInmuebles(){
 export async function fetchAgregarinmueble(inmueble){
     
     try {
-        const respuesta= await fetch(import.meta.env.VITE_API_URL,{
+        const respuesta= await fetch(`${import.meta.env.VITE_API_URL}/fetchInmuebles`,{
             method: 'POST',
             body: JSON.stringify(inmueble),
             headers: {
                 'Content-Type': 'application/json'
                 }
         })
-        await respuesta.json()
+        const resultado= await respuesta.json()
     } catch (error) {
         console.log(error)
     }
@@ -23,7 +23,7 @@ export async function fetchAgregarinmueble(inmueble){
 
 export async function fetchEditarInmueble({ titulo,descripcion,precio,direccion,ambientes,totales,cubiertos,estrellas,id}){
     try {
-        const respuesta = await fetch(`${import.meta.env.VITE_API_URL}/${id}`,{
+        const respuesta = await fetch(`${import.meta.env.VITE_API_URL}/fetchInmuebles/${id}`,{
             method: 'PUT',
             body: JSON.stringify({
                 titulo,
@@ -40,7 +40,7 @@ export async function fetchEditarInmueble({ titulo,descripcion,precio,direccion,
                 'Content-Type': 'application/json'
             }
         })
-        await respuesta.json()
+        const resultado= await respuesta.json()
     } catch (error) {
         console.log(error)
     }
@@ -48,11 +48,24 @@ export async function fetchEditarInmueble({ titulo,descripcion,precio,direccion,
 
 export async function fetchEliminarInmueble(id){
     try {
-        const respuesta = await fetch(`${import.meta.env.VITE_API_URL}/${id}`,{
-            method: 'DELETE'
-            })
-            await respuesta.json()
+        if (id) {
+            const respuesta = await fetch(`${import.meta.env.VITE_API_URL}/fetchInmuebles/${id}`,{
+                method: 'DELETE'
+                })
+                const resultado=  await respuesta.json()
+                const result = [{status:200, resultado}]
+                console.log(result)
+                return result
+        } else {
+        const result = [{status:404, error:"No me pasaste el id"}]
+        return result
+
+        }
+    
     } catch (error) {
         console.log(error)
+        const result = [{status:404, error:"Hubo un error en la elimiacion"}]
+        return result
+
     }
 }
